requirements.txt
django
djangorestframework
django-cors-headers
channels
channels-redis
